Move User Model into Models directory
