+++ app/Http/Controllers/Auth/ForgotPasswordController.php
@@ -2,7 +2,7 @@
 namespace App\Http\Controllers\Auth;
 use App\Http\Controllers\Controller;
 use App\Transformers\Json;
-use App\User;
+use App\Models\User;
 use Illuminate\Foundation\Auth\SendsPasswordResetEmails;
 use Illuminate\Http\Request;
 class ForgotPasswordController extends Controller
