+++ app/Http/Controllers/Auth/LoginController.php
@@ -6,7 +6,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use JWTAuth;
 use Tymon\JWTAuth\Exceptions\JWTException;
-use App\User;
+use App\Models\User;
 use Socialite;
 
 class LoginController extends Controller
